Attach owner to each survey
created a `for loop` in the getUserSurveySuccess Ui so that if the
owner'd id = user id, it should send the survey in an empty object and
array.

To do: Figure out why the handlebars are not showing the user survey.
Something to do with the way the file is attached or how the
handlebars are parsing it.